fix(dtype): explicit casts at 6 scatter sites to silence FutureWarning
JAX scatter (`array.at[idx].set(value)`) now warns and will soon error when
`value.dtype` is wider than `array.dtype`. Fixes 6 sites where implicit
downcasting was occurring during step tracing:

- `dungeon/spawning.py:2316-2319` — `_ATK_DICE_N`/`_ATK_DICE_S` (int16) into
  `attack_dice_n`/`attack_dice_sides` (int8) for oroom monster scanbody.
- `subsystems/status_effects.py:1032` — `new_fainting_turns` derived from
  `jax.random.randint` (int64 under JAX_ENABLE_X64) into `timed_statuses[]`
  (int32) for FAINTING_TURNS.
- `subsystems/items_wands.py:1566-1567` — `new_hp`/`new_hp_max` into
  `mon_hp`/`mon_hp_max` (int32) for wand-draining effect.
- `subsystems/magic.py:1289,1301` — `jnp.int32(MoveStrategy.*)` into
  `mstrategy` (int8) for sleep + confuse monster effects.
- `subsystems/quest.py:473-474` — `atk_n`/`atk_s` (int16) into
  `attack_dice_n`/`attack_dice_sides` (int8) for quest leader spawn.

Each fix uses explicit `.astype()` to the target dtype. All values fit
trivially in the target dtype (monster attack dice ≤ 8d8, faint dur 1-10,
mstrategy 0-30) so no information loss. No byteparity impact.

+++ Nethax/nethax/dungeon/spawning.py
@@ -2182,8 +2182,8 @@ def _populate_oroom_single(
         alive=mai.alive.at[slot].set(jnp.bool_(True)),
         ac=mai.ac.at[slot].set(_BASE_AC[tid]),
         is_large=mai.is_large.at[slot].set(_IS_LARGE[tid]),
-        attack_dice_n=mai.attack_dice_n.at[slot].set(_ATK_DICE_N[tid]),
-        attack_dice_sides=mai.attack_dice_sides.at[slot].set(_ATK_DICE_S[tid]),
+        attack_dice_n=mai.attack_dice_n.at[slot].set(_ATK_DICE_N[tid].astype(jnp.int8)),
+        attack_dice_sides=mai.attack_dice_sides.at[slot].set(_ATK_DICE_S[tid].astype(jnp.int8)),
         mstrategy=mai.mstrategy.at[slot].set(jnp.int8(0)),
         entry_idx=mai.entry_idx.at[slot].set(tid.astype(jnp.int16)),
         peaceful=mai.peaceful.at[slot].set(peace_bit),
@@ -2316,8 +2316,8 @@ def spawn_oroom_monster_scanbody(
                 alive=mai_g.alive.at[slot_g].set(jnp.bool_(True)),
                 ac=mai_g.ac.at[slot_g].set(_BASE_AC[tid]),
                 is_large=mai_g.is_large.at[slot_g].set(_IS_LARGE[tid]),
-                attack_dice_n=mai_g.attack_dice_n.at[slot_g].set(_ATK_DICE_N[tid]),
-                attack_dice_sides=mai_g.attack_dice_sides.at[slot_g].set(_ATK_DICE_S[tid]),
+                attack_dice_n=mai_g.attack_dice_n.at[slot_g].set(_ATK_DICE_N[tid].astype(jnp.int8)),
+                attack_dice_sides=mai_g.attack_dice_sides.at[slot_g].set(_ATK_DICE_S[tid].astype(jnp.int8)),
                 mstrategy=mai_g.mstrategy.at[slot_g].set(jnp.int8(0)),
                 entry_idx=mai_g.entry_idx.at[slot_g].set(tid.astype(jnp.int16)),
                 peaceful=mai_g.peaceful.at[slot_g].set(peace_bit),
@@ -2522,8 +2522,8 @@ def populate_level_with_monsters(
         new_alive     = mai_carry.alive.at[i].set(jnp.bool_(True))
         new_ac        = mai_carry.ac.at[i].set(_BASE_AC[type_id])
         new_is_large  = mai_carry.is_large.at[i].set(_IS_LARGE[type_id])
-        new_atk_n     = mai_carry.attack_dice_n.at[i].set(_ATK_DICE_N[type_id])
-        new_atk_s     = mai_carry.attack_dice_sides.at[i].set(_ATK_DICE_S[type_id])
+        new_atk_n     = mai_carry.attack_dice_n.at[i].set(_ATK_DICE_N[type_id].astype(jnp.int8))
+        new_atk_s     = mai_carry.attack_dice_sides.at[i].set(_ATK_DICE_S[type_id].astype(jnp.int8))
         new_strategy  = mai_carry.mstrategy.at[i].set(jnp.int8(0))  # NONE until awakened
         new_entry     = mai_carry.entry_idx.at[i].set(type_id.astype(jnp.int16))
         # Vendor mtmp->movement starts at 0 (calloc default — makemon.c has
